Major refactor and version update
Dragged everything forward to CDK 2.0, and did a big refactor to get
everything cleaner and easier to work with. No significant changes were
made to the implementation other than moving it from an existing VPC
(where it looks for the default VPC), to a model where a new VPC is
created by the template.