Added file asset mechanism and Powerlevel10k support
Dragged in PowerLevel10k for Zsh theme and functionality, which is cool.
It did however push my userdata over the size limit with the appraoch I
was using to write files to disk, so I put a file map in to the config
and pushed the zshrc file out to there. This will make it much easier
to add local files to the instance (noting they go in after the commands
have all fininshed running).