Added loadEnv function to zshrc file
This will look in the current directory and load the .env file in to
the current session. Each line in the .env should be stored as

VAR=VALUE

Makes it perfect to store variables in to that are proejct specific.
Just be sure to not commit it if you don't need it in source control
(especially if it has secrets).